chore : Update constants
