+++ src/modules/auth/pages/ChangePasswordPage.vue
@@ -7,7 +7,7 @@
 </template>
 
 <script setup lang="ts">
-  import { inputForms } from 'src/constants/form.constants';
+  import { inputForms } from 'auth/constants/form';
   import type { BodyForm } from 'shared/models/form';
   import type { ChangePasswordBody } from 'auth/models/fusion-auth.models';
   import { useAuthStore } from 'src/modules/auth/stores/auth-store';
