+++ src/modules/shared/helpers/router.ts
@@ -1,8 +1,8 @@
-import { useAuthStore } from 'src/modules/auth/stores/auth-store';
+import { useAuthStore } from 'auth/stores/auth-store';
 import type { NavigationGuardNext, RouteLocationNormalizedGeneric } from 'vue-router';
 import type { AxiosError } from 'axios';
 import { UserService } from 'shared/services/user';
-import { decodeJwt } from 'src/utils/auth.utils';
+import { decodeJwt } from 'auth/helpers/jwt';
 import { AUTH_ROUTES } from 'auth/routes/route-names';
 import { getSessionStorageValues } from 'auth/helpers/sessionStorage';
